\name{polio}
\alias{polio}
\docType{data}
\title{
Polio Time Series 
}
\description{
Time series of Polio incidences in USA from 1970 to 1983.
}
\usage{data(polio)}
\format{
Time series of monthly Polio cases in USA from January 1970 to December 1983.
}
\source{
Zeger, S.L. (1988). A regression model for time series of counts. \emph{Biometrika} \bold{75}, 822--835.
}
\examples{
data(polio, package = "lacm")
}
\keyword{datasets}
